Relocate feature components
